+++ src/Analyzers/CSharp/DiagnosticIdentifiers.Generated.cs
@@ -214,5 +214,6 @@ namespace Roslynator.CSharp
         public const string SimplifyArgumentNullCheck = "RCS1255";
         public const string InvalidArgumentNullCheck = "RCS1256";
         public const string UseEnumFieldExplicitly = "RCS1257";
+        public const string UnnecessaryEnumFlag = "RCS1258";
     }
 }